wildfire data parser (just dbf file)
assumes you have the file zipped and in the same directory. Next goal is to get this stored into a dataframe with pandas. I can't make heads or tails of the parameters it describes, besides a few. I have no idea how this is going to be converted into a shape file